+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Revert back to `all` operator for title (All) index. Fixes UIIN-2274.
 * Warn on `sessionStorage` errors instead of swallowing them.
 * ISRI: Update the Settings screen to allow multiple job profiles: Create/Edit. Refs UIIN-2248.
+* ISRI: Update the Settings screen to allow multiple job profiles: View. Refs UIIN-2249.
 
 ## [9.2.0](https://github.com/folio-org/ui-inventory/tree/v9.2.0) (2022-10-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.1.0...v9.2.0)